Added RemoteableTimeSignature and ScaleInformation for LiveSet and MidiClip

+++ service/internal/sourcer/abletonv5/xml_base_values.go
@@ -5,7 +5,7 @@ type XmlStringValue struct {
 }
 
 type XmlIntValue struct {
-	Value float64 `xml:"Value,attr"`
+	Value int64 `xml:"Value,attr"`
 }
 
 type XmlFloatValue struct {
